second change;

+++ spring_boot_backend_HotelResProj/src/main/java/com/app/repository/BookingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.app.entity.Booking;
 @Repository
-public interface BookingRepository extends JpaRepository<Booking, Long> {
+public interface BookingRepository extends JpaRepository<Booking, Integer>{
 
 }
